:ref: moving common servlet filters, shared by job-engine and rest-api webapps, to a new module under the newly created commons-rest
