+++ pom.xml
@@ -979,6 +979,11 @@
             </dependency>
             <dependency>
                 <groupId>org.eclipse.kapua</groupId>
+                <artifactId>kapua-commons-rest-filters</artifactId>
+                <version>2.0.0-SNAPSHOT</version>
+            </dependency>
+            <dependency>
+                <groupId>org.eclipse.kapua</groupId>
                 <artifactId>kapua-scheduler-api</artifactId>
                 <version>${project.version}</version>
             </dependency>
